+++ R/barplot.R
@@ -58,8 +58,8 @@ barplot.enrichResult <- function(height, x="Count", color='p.adjust',
     if(colorBy %in% colnames(df)) {
         p <- ggplot(df, aes_string(x = x, y = "Description", fill = colorBy)) +
             theme_dose(font.size) +
-            scale_fill_continuous(low="red", high="blue", name = color,
-                                  guide=guide_colorbar(reverse=TRUE))
+            # scale_fill_continuous(name = color) + 
+            set_enrichplot_color(type = "fill", name = color)
     } else {
         p <- ggplot(df, aes_string(x = x, y = "Description",
                                    fill = "Description")) +