+++ R/dotplot.R
@@ -35,7 +35,8 @@ setMethod("dotplot", signature(object = "gseaResult"),
                 
                 if (NES) {
                     p <- suppressMessages(p + aes_(color=~NES) + 
-                        scale_color_continuous(low="blue", high="red", name = "NES")
+                        # scale_fill_continuous(name = "NES") +
+                        set_enrichplot_color(type = "fill", name = "NES")
                     )
                 }
                 return(p)
@@ -106,7 +107,8 @@ setMethod("dotplot", signature(object = "gseaResultList"),
               
               if (NES) {
                   p <- suppressMessages(p + aes_(color=~NES) + 
-                      scale_color_continuous(low="blue", high="red", name = "NES")
+                      # scale_fill_continuous(name = "NES") + 
+                      set_enrichplot_color(type = "fill", name = "NES")
                   )
               }
               return(p)
@@ -196,15 +198,13 @@ dotplot.enrichResult <- function(object, x = "geneRatio", color = "p.adjust",
 
     df$Description <- factor(df$Description,
                           levels=rev(unique(df$Description[idx])))
-    ggplot(df, aes_string(x=x, y="Description", size=size, color=colorBy)) +
-        geom_point() +
-        scale_color_continuous(low="red", high="blue", name = color,
-            guide=guide_colorbar(reverse=TRUE)) +
+    ggplot(df, aes_string(x=x, y="Description", size=size, fill=colorBy)) +
+        geom_point(shape = 21) +
+        # scale_fill_continuous(name = color) +
+        set_enrichplot_color(type = "fill", name = color) + 
         scale_y_discrete(labels = label_func) +
         ylab(NULL) + ggtitle(title) + theme_dose(font.size) +
-        scale_size(range=c(3, 8)) +
-        guides(size  = guide_legend(order = 1),
-               color = guide_colorbar(order = 2))
+        scale_size(range=c(3, 8))
 }
 
 
@@ -270,17 +270,14 @@ dotplot.compareClusterResult <- function(object, x= "Cluster", colorBy="p.adjust
         require(ggstar, character.only=TRUE)
         # p <- p + ggsymbol::geom_symbol(aes_string(symbol = "Cluster", fill = colorBy)) +
         p <- p + ggstar::geom_star(aes_string(starshape="Cluster", fill=colorBy)) +
-            scale_fill_continuous(low="red", high="blue", guide=guide_colorbar(reverse=TRUE))
+            set_enrichplot_color(type = "fill")
     }  else {
-        p <- p +  geom_point(aes_string(color = colorBy))
+        p <- p +  geom_point(shape = 21, aes_string(color = colorBy))
     }
 
-    p <- p + scale_color_continuous(low="red", high="blue",
-                    guide=guide_colorbar(reverse=TRUE)) +
+    p <- p + set_enrichplot_color(type = "fill") +
         ylab(NULL) + ggtitle(title) + DOSE::theme_dose(font.size) +
-        scale_size_continuous(range=c(3, 8)) +
-        guides(size  = guide_legend(order = 1),
-                color = guide_colorbar(order = 2))
+        scale_size_continuous(range=c(3, 8))
 
 
     if (!is.null(facet)) {