+++ R/treeplot.R
@@ -547,10 +547,9 @@ treeplot.compareClusterResult <-  function(x,
         p <- p + ggnewscale::new_scale_fill() # +
             # coord_equal(xlim = xlim)
         p <- ggtree::gheatmap(p, ID_Cluster_mat) + 
-            scale_fill_continuous(low="red", high="blue", 
-                                  guide = guide_colorbar(reverse=TRUE),
-                                  trans = "log10",
-                                  name = color) 
+            # scale_fill_continuous(trans = "log10", name = color) +
+            set_enrichplot_color(type = "fill", trans = "log10", name = color)
+            
     }
 
     if (geneClusterPanel == "dotplot") {
@@ -566,11 +565,8 @@ treeplot.compareClusterResult <-  function(x,
                                      size = "Count", color = color),
                        pwidth = 0.5, offset = -0.2,
                        axis.params = list(axis = "x", text.size = 3, line.alpha = 0)) +
-            scale_colour_continuous(low="red", high="blue", 
-                                  guide=guide_colorbar(reverse=TRUE),
-                                  trans = "log10",
-                                  name = color) # + 
-            # coord_equal(xlim = xlim) 
+            # scale_colour_continuous(trans = "log10", name = color) + 
+            set_enrichplot_color(trans = "log10", name = color)
             
     }
     p + ggtree::hexpand(ratio = hexpand) + coord_equal()
@@ -727,8 +723,8 @@ group_tree <- function(hc, clus, d, offset_tiplab, nWords,
     if (add_tippoint) {
         p <- p + ggnewscale::new_scale_colour() +
             geom_tippoint(aes(color = color, size = count)) + 
-            scale_colour_continuous(low="red", high="blue", name = color, 
-                guide = guide_colorbar(reverse = TRUE))
+            # scale_colour_continuous(name = color)+
+            set_enrichplot_color(name = color)
     }
     ## add tiplab 
     p <- p + geom_tiplab(offset = offset_tiplab, hjust = 0,