+++ NEWS.md
@@ -1,5 +1,6 @@
 # enrichplot 1.21.2.002
 
++ change default color: from c("red", "blue") to c("#e06663", "#327eba") (2023_9_13, Wed)
 + `set_enrichplot_color()`, a helper function to set colors (2023-09-13, Wed)
 + use `check_installed()` to check package dependency (2023-09-08, Fri, #254)